auth0 goal integration

+++ src/main/java/br/edu/insper/goals/GoalController.java
@@ -2,6 +2,8 @@ package br.edu.insper.goals;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.security.oauth2.jwt.Jwt;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.server.ResponseStatusException;
 
@@ -12,8 +14,13 @@ public class GoalController {
     private GoalService goalService;
 
     @GetMapping("/goal/{id}")
-    public Goal getGoal(@PathVariable int id){
+    public Goal getGoal(@AuthenticationPrincipal Jwt jwt, @PathVariable int id){
         Goal goal = goalService.getGoal(id);
+        String email = jwt.getClaimAsString("https://stocks-insper.com/email");
+
+        if (email.isEmpty()) {
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED);
+        }
         if (goal.getId() == 0) {
             throw new ResponseStatusException(HttpStatus.NOT_FOUND);
         }
@@ -21,13 +28,23 @@ public class GoalController {
     }
 
     @PostMapping("/goal")
-    public Goal postGoal(@RequestBody Goal goal) {
+    public Goal postGoal(@AuthenticationPrincipal Jwt jwt, @RequestBody Goal goal) {
+        String email = jwt.getClaimAsString("https://stocks-insper.com/email");
+
+        if (email.isEmpty()) {
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED);
+        }
         goalService.postGoal(goal);
         return goal;
     }
 
     @PutMapping("/goal/{id}")
-    public Goal putGoal(@PathVariable int id, @RequestBody Goal updatedGoal) {
+    public Goal putGoal(@AuthenticationPrincipal Jwt jwt, @PathVariable int id, @RequestBody Goal updatedGoal) {
+        String email = jwt.getClaimAsString("https://stocks-insper.com/email");
+
+        if (email.isEmpty()) {
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED);
+        }
         if (updatedGoal.getDescription() == null) {
             throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
         }
@@ -36,7 +53,13 @@ public class GoalController {
 
     @DeleteMapping("/goal/{id}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
-    public Goal deleteGoal(@PathVariable int id) {
+    public Goal deleteGoal(@AuthenticationPrincipal Jwt jwt, @PathVariable int id) {
+        String email = jwt.getClaimAsString("https://stocks-insper.com/email");
+
+        if (email.isEmpty()) {
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED);
+        }
+
         return goalService.removeGoal(id);
     }
 }
